+++ src/lib/pages/deploy-script/components/index.ts
@@ -1,0 +1,3 @@
+export * from "./Footer";
+export * from "./ScriptInput";
+export * from "./UploadScriptCard";